Bump to 0.2.0
This minor-level bump switches the underlying OAuth libraries. Although
this change should be transparent to end users, if an issue were to
arise from this change it would make that version of the tool unusuable.
As a result, we choose to make this a minor-level change so it will be
easier for users to recognize that a change occurred and adjust
accordingly.