Throw error on cell formula error
If a cell in a Google Sheet has a formula error (for example, the cell
has the formula '=1/0'), then running fetch_data() will fail as the cell
in question has no data: it only has the error.

Previous to this commit we didn't surface this information, leading to a
confusing stack trace that didn't indicate that the culprit was a cell
formula error (see issue #11 as a motivating example for this change).

This commit raises an exception when we have a cell formula error,
indicating the cell that errored and also passing along Google Sheets'
formula error for that cell.

Note that this approach throws an error at the first formula error that is
encountered; any other formula errors that exist will not be reported as
the library didn't reach them before exiting execution.